Cost function parameter to study loss and alpha
Adding cost function parameter display to study if Cost
function is indeed converging. We also identify how many
iterations were required for the convergence.
Iterations are based on how small the learning rate is.
I use the insight that the cost function will not reduce
once we reach global optima.

Signed-off-by: Pankaj Dabade <dabade@usc.edu>